requirements.txt
anthropic>=0.40.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
numpy>=1.24.0
